+++ src/Command/ConfigCommandProvider.php
@@ -16,6 +16,7 @@ final class ConfigCommandProvider implements CommandProvider
         return [
             new CopyCommand(),
             new RebuildCommand(),
+            new InfoCommand(),
         ];
     }
 }