+++ tests/Command/ConfigCommandProviderTest.php
@@ -9,6 +9,7 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Command\Command;
 use Yiisoft\Config\Command\ConfigCommandProvider;
 use Yiisoft\Config\Command\CopyCommand;
+use Yiisoft\Config\Command\InfoCommand;
 use Yiisoft\Config\Command\RebuildCommand;
 
 final class ConfigCommandProviderTest extends TestCase
@@ -17,12 +18,15 @@ final class ConfigCommandProviderTest extends TestCase
     {
         $provider = new ConfigCommandProvider();
 
-        $this->assertCount(2, $provider->getCommands());
+        $this->assertCount(3, $provider->getCommands());
         $this->assertInstanceOf(BaseCommand::class, $provider->getCommands()[0]);
         $this->assertInstanceOf(BaseCommand::class, $provider->getCommands()[1]);
+        $this->assertInstanceOf(BaseCommand::class, $provider->getCommands()[2]);
         $this->assertInstanceOf(Command::class, $provider->getCommands()[0]);
         $this->assertInstanceOf(Command::class, $provider->getCommands()[1]);
+        $this->assertInstanceOf(Command::class, $provider->getCommands()[2]);
         $this->assertEquals($provider->getCommands()[0], new CopyCommand());
         $this->assertEquals($provider->getCommands()[1], new RebuildCommand());
+        $this->assertEquals($provider->getCommands()[2], new InfoCommand());
     }
 }