+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## 1.5.1 under development
 
+- New #175: Add `yii-config-info` composer command (@vjik)
 - New #173: Allow to use option "config-plugin-file" in packages (@vjik) 
 - Enh #172, #173: Refactoring: extract config settings reader to separate class (@vjik)
+- Chg #175: Raise minimum Composer version to 2.3 (@vjik)
+- Enh #175: Minor refactoring of internal classes `Options` and `ProcessHelper` (@vjik)
 
 ## 1.5.0 December 25, 2023